refactor(io): move question related code to 'ask' module
