+++ mcmd/io/io.py
@@ -1,10 +1,9 @@
-from PyInquirer import prompt
 from colorama import Fore, Style
 from halo import Halo
 
 import mcmd.config.config as config
-from mcmd.io.logging import get_logger
 from mcmd.io.kbhit import KBHit
+from mcmd.io.logging import get_logger
 
 log = get_logger()
 
@@ -76,81 +75,6 @@ def debug(message):
     log.debug('  ' + message)
 
 
-def multi_choice(message, choices):
-    if spinner:
-        spinner.stop_and_persist()
-
-    message = {
-        'type': 'rawlist',
-        'name': 'answer',
-        'message': message,
-        'choices': choices
-    }
-
-    return _handle_question(message)
-
-
-def checkbox(message, choices):
-    if spinner:
-        spinner.stop_and_persist()
-
-    checks = [{'name': choice, 'value': idx} for idx, choice in enumerate(choices)]
-
-    message = {
-        'type': 'checkbox',
-        'name': 'answer',
-        'message': message,
-        'choices': checks,
-        'validate': lambda answer: 'You must choose at least one option.' if len(answer) == 0 else True
-    }
-
-    answer_ids = _handle_question(message)
-    return [choices[idx] for idx in answer_ids]
-
-
-def input_(message, required=False):
-    if spinner:
-        spinner.stop_and_persist()
-
-    message = {
-        'type': 'input',
-        'name': 'answer',
-        'message': message,
-    }
-
-    if required:
-        message['validate'] = lambda answer: "This field can't be empty." if len(answer) == 0 else True
-
-    return _handle_question(message)
-
-
-def password(message):
-    if spinner:
-        spinner.warn()
-
-    message = {
-        'type': 'password',
-        'name': 'answer',
-        'message': message,
-    }
-
-    return _handle_question(message)
-
-
-def confirm(message):
-    if spinner:
-        spinner.stop_and_persist()
-
-    message = {
-        'type': 'confirm',
-        'default': False,
-        'name': 'answer',
-        'message': message
-    }
-
-    return _handle_question(message)
-
-
 def wait_for_enter():
     """Waits until the user presses enter. Non-blocking: the program can still be interrupted and closed."""
     kb = KBHit()
@@ -184,14 +108,3 @@ def bold(string):
 
 def dim(string):
     return Style.DIM + string + Style.RESET_ALL
-
-
-def _handle_question(question):
-    answer = prompt([question])
-    if not answer:
-        # empty result means that PyInquirer caught an InterruptException
-        exit(0)
-    else:
-        if spinner:
-            spinner.start()
-        return answer['answer']
