+++ tests/integration/commands/test_delete.py
@@ -7,7 +7,7 @@ from tests.integration.utils import run_commander, entity_type_exists, entity_is
 
 
 @pytest.mark.integration
-@patch('mcmd.io.io.confirm')
+@patch('mcmd.io.ask.confirm')
 def test_delete_entity(are_you_sure, session, entity_type):
     are_you_sure.return_value = True
     run_commander('delete --entity-type {}'.format(entity_type))
@@ -16,7 +16,7 @@ def test_delete_entity(are_you_sure, session, entity_type):
 
 
 @pytest.mark.integration
-@patch('mcmd.io.io.confirm')
+@patch('mcmd.io.ask.confirm')
 def test_delete_entity_cancel(are_you_sure, session, entity_type):
     are_you_sure.return_value = False
     run_commander('delete --entity-type {}'.format(entity_type))
@@ -25,7 +25,7 @@ def test_delete_entity_cancel(are_you_sure, session, entity_type):
 
 
 @pytest.mark.integration
-@patch('mcmd.io.io.confirm')
+@patch('mcmd.io.ask.confirm')
 def test_delete_force(are_you_sure, session, package):
     run_commander('delete --force --package {}'.format(package))
 
@@ -41,7 +41,7 @@ def test_delete_entity_data(session, entity_type):
 
 
 @pytest.mark.integration
-@patch('mcmd.io.io.confirm')
+@patch('mcmd.io.ask.confirm')
 def test_delete_entity_attribute(are_you_sure, session, entity_type):
     are_you_sure.return_value = True
     run_commander('delete --entity-type {} --attribute firstName'.format(entity_type))
