+++ tests/integration/commands/test_disable_rls.py
@@ -7,7 +7,7 @@ from tests.integration.utils import run_commander
 
 
 @pytest.mark.integration
-@patch('mcmd.io.io.confirm')
+@patch('mcmd.io.ask.confirm')
 def test_disable_rls(are_you_sure, session, entity_type):
     are_you_sure.return_value = True
     run_commander('enable rls {}'.format(entity_type))
@@ -17,7 +17,7 @@ def test_disable_rls(are_you_sure, session, entity_type):
 
 
 @pytest.mark.integration
-@patch('mcmd.io.io.confirm')
+@patch('mcmd.io.ask.confirm')
 def test_disable_rls_cancel(are_you_sure, session, entity_type):
     are_you_sure.return_value = False
     run_commander('enable rls {}'.format(entity_type))
