+++ tests/unit/commands/import_test.py
@@ -9,7 +9,7 @@ from mcmd.github.client import Attachment
 
 @pytest.mark.unit
 class ImportMethodsTest(unittest.TestCase):
-    @patch('mcmd.io.io.multi_choice')
+    @patch('mcmd.io.ask.multi_choice')
     def test_choose_attachments(self, multi_choice):
         multi_choice.return_value = 'other_file.xlsx'
         a1 = Attachment('url/1234/file.xlsx')
@@ -17,7 +17,7 @@ class ImportMethodsTest(unittest.TestCase):
         ret = import_._choose_attachment([a1, a2])
         self.assertEqual(ret, a2)
 
-    @patch('mcmd.io.io.multi_choice')
+    @patch('mcmd.io.ask.multi_choice')
     def test_choose_attachments_with_duplicates(self, multi_choice):
         multi_choice.return_value = '1234/file.xlsx'
         a1 = Attachment('url/1234/file.xlsx')
